data-platform-mlops/data_platform_mlops/requirements.txt
mlflow==2.8.1
numpy>=1.23.0
pandas>=1.4.3
scikit-learn>=1.1.1
matplotlib>=3.5.2
Jinja2==3.0.3
pyspark~=3.3.0
pytz~=2022.2.1
databricks-feature-store~=0.16.2